JENKINS-17910 - groovy.lang.MissingMethodException use proper class instead of inner class
